make `ConditionEvents` an interface
This way it is easier to maintain it in the future since we can freely adapt / optimize the implementation.
