+++ archunit/src/main/java/com/tngtech/archunit/lang/conditions/ContainsOnlyCondition.java
@@ -34,7 +34,7 @@ class ContainsOnlyCondition<T> extends ArchCondition<Collection<? extends T>> {
 
     @Override
     public void check(Collection<? extends T> collection, ConditionEvents events) {
-        ConditionEvents subEvents = new ConditionEvents();
+        ConditionEvents subEvents = ConditionEvents.Factory.create();
         for (T item : collection) {
             condition.check(item, subEvents);
         }
