+++ archunit/src/test/java/com/tngtech/archunit/lang/EvaluationResultTest.java
@@ -81,7 +81,7 @@ public class EvaluationResultTest {
     }
 
     private ConditionEvents events(ConditionEvent... events) {
-        ConditionEvents result = new ConditionEvents();
+        ConditionEvents result = ConditionEvents.Factory.create();
         for (ConditionEvent event : events) {
             result.add(event);
         }
